moved space to abstract and concrete
