+++ mesa_frames/concrete/model.py
@@ -4,8 +4,8 @@ from typing import TYPE_CHECKING
 import numpy as np
 from typing_extensions import Any
 
+from mesa_frames.abstract.space import MultiSpaceDF, SpaceDF
 from mesa_frames.concrete.agents import AgentsDF
-from mesa_frames.concrete.space import SpaceDF
 
 if TYPE_CHECKING:
     from mesa_frames.abstract.agents import AgentSetDF
@@ -60,7 +60,7 @@ class ModelDF:
     _seed: int | Sequence[int]
     running: bool
     _agents: AgentsDF
-    _space: SpaceDF | None
+    _space: MultiSpaceDF
 
     def __new__(
         cls, seed: int | Sequence[int] | None = None, *args: Any, **kwargs: Any
@@ -79,6 +79,7 @@ class ModelDF:
         self.schedule = None
         self.current_id = 0
         self._agents = AgentsDF(self)
+        self._space = MultiSpaceDF(self)
 
     def get_agents_of_type(self, agent_type: type) -> "AgentSetDF":
         """Retrieve the AgentSetDF of a specified type.
@@ -160,4 +161,4 @@ class ModelDF:
 
     @space.setter
     def space(self, space: SpaceDF) -> None:
-        self._space = space
+        self._space.add(space)
